third_party/WebKit/LayoutTests/inspector/debugger/script-snippet-model.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script>
function test()
{
    function resetSnippetsSettings()
    {
        WebInspector.scriptSnippetModel._snippetStorage.reset();
        WebInspector.scriptSnippetModel._lastSnippetEvaluationIndexSetting.set(0);
        WebInspector.scriptSnippetModel.reset();
    }

    var workspace = new WebInspector.Workspace();
    var snippetWorkspaceProvider = new WebInspector.SimpleWorkspaceProvider(workspace, WebInspector.projectTypes.Snippets);
    var workspaceController = new WebInspector.WorkspaceController(workspace);
    WebInspector.scriptSnippetModel = new WebInspector.ScriptSnippetModel(workspace, snippetWorkspaceProvider);
    InspectorTest.runDebuggerTestSuite([
        function testCreateAndRename(next)
        {
            function renameSnippetAndCheckWorkspace(uiSourceCode, snippetName)
            {
                InspectorTest.addResult("Renaming snippet to '" + snippetName + "' ...");
                uiSourceCode.rename(snippetName, renameCallback);

                function renameCallback(success)
                {
                    if (success)
                        InspectorTest.addResult("Snippet renamed successfully.");
                    else
                        InspectorTest.addResult("Snippet was not renamed.");
                }
                InspectorTest.addResult("UISourceCode name is '" + uiSourceCode.name() + "' now.");
                InspectorTest.addResult("Number of uiSourceCodes in workspace: " + workspace.uiSourceCodes().length);
            }

            resetSnippetsSettings();
            var uiSourceCode1 = WebInspector.scriptSnippetModel.createScriptSnippet();
            InspectorTest.addResult("Snippet1 created.");
            var uiSourceCode2 = WebInspector.scriptSnippetModel.createScriptSnippet();
            InspectorTest.addResult("Snippet2 created.");
            renameSnippetAndCheckWorkspace(uiSourceCode1, "foo");
            renameSnippetAndCheckWorkspace(uiSourceCode1, "   ");
            renameSnippetAndCheckWorkspace(uiSourceCode1, " bar ");
            renameSnippetAndCheckWorkspace(uiSourceCode1, "foo");
            renameSnippetAndCheckWorkspace(uiSourceCode2, "bar");
            renameSnippetAndCheckWorkspace(uiSourceCode2, "foo");
            next();
        },

        function testEvaluate(next)
        {
            resetSnippetsSettings();

            var uiSourceCode1 = WebInspector.scriptSnippetModel.createScriptSnippet();
            uiSourceCode1.rename("Snippet1", function() { });
            var content = "";
            content += "// This snippet does nothing.\n";
            content += "var i = 2+2;\n";
            uiSourceCode1.setWorkingCopy(content);

            var uiSourceCode2 = WebInspector.scriptSnippetModel.createScriptSnippet();
            uiSourceCode2.rename("Snippet2", function() { });
            content = "";
            content += "// This snippet creates a function that does nothing and returns it.\n";
            content += "function doesNothing() {\n";
            content += "    var  i = 2+2;\n";
            content += "};\n";
            content += "doesNothing;\n";
            uiSourceCode2.setWorkingCopy(content);

            function evaluateSnippetAndDumpEvaluationDetails(uiSourceCode, callback)
            {
                InspectorTest.addSniffer(WebInspector.SnippetScriptMapping.prototype, "addScript", dumpScript);
                // JSC does not support compile-and-run for snippet evaluation, so different platforms have slightly different codeflow here, hence two sniffers.
                InspectorTest.addSniffer(WebInspector.ConsoleView.prototype, "_printResult", dumpResult);
                InspectorTest.addSniffer(WebInspector.ScriptSnippetModel.prototype, "_printRunScriptResult", dumpResult);
                WebInspector.scriptSnippetModel.evaluateScriptSnippet(uiSourceCode);
                var evaluationSourceURL = WebInspector.scriptSnippetModel._evaluationSourceURL(uiSourceCode);
                var snippetId = WebInspector.scriptSnippetModel._snippetIdForUISourceCode.get(uiSourceCode);
                InspectorTest.addResult("Last evaluation source url for snippet: " + evaluationSourceURL);
                InspectorTest.assertEquals(snippetId, WebInspector.scriptSnippetModel._snippetIdForSourceURL(evaluationSourceURL), "Snippet can not be identified by its evaluation sourceURL.");

                function dumpScript(script)
                {
                    InspectorTest.addResult("Snippet script added, sourceURL = " + script.sourceURL);
                }

                function dumpResult(result, wasThrown)
                {
                    InspectorTest.addResult("Snippet execution result: " + result.description);
                    callback();
                }
            }

            evaluateSnippetAndDumpEvaluationDetails(uiSourceCode1, step2);

            function step2()
            {
                evaluateSnippetAndDumpEvaluationDetails(uiSourceCode2, step3);
            }

            function step3()
            {
                evaluateSnippetAndDumpEvaluationDetails(uiSourceCode1, next);
            }
        }
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests script snippet model.</p>
</body>
</html>
